runtime: fix warnings in GCC

+++ runtime/include/alaska/HandleTable.hpp
@@ -22,7 +22,7 @@
 
 namespace alaska {
 
-  using slabidx_t = uint64_t;
+  using slabidx_t = size_t;
 
   class HandleTable;
   class HandleSlabQueue;
